PHP8.1 Deprecated: addslashes(): Passing null to parameter #1 ($string) of type string is deprecated. addslashes 함수의 문자타입의 첫번째 파라미터에 null을 전달하는 것은 더 이상 사용되지 않습니다. 해결

+++ lib/PHPExcel/Cell.php
@@ -842,7 +842,7 @@ class PHPExcel_Cell
             if ($pColumnIndex < 26) {
                 $_indexCache[$pColumnIndex] = chr(65 + $pColumnIndex);
             } elseif ($pColumnIndex < 702) {
-                $_indexCache[$pColumnIndex] = chr(64 + ($pColumnIndex / 26)) .
+                $_indexCache[$pColumnIndex] = chr(64 + (int)($pColumnIndex / 26)) .
                                               chr(65 + $pColumnIndex % 26);
             } else {
                 $_indexCache[$pColumnIndex] = chr(64 + (($pColumnIndex - 26) / 676)) .
